Cleanup from faulty merge
NB! The Merge accidentally duped the "utiles" which made everything go all whacky for me, as the headless chrome tried to direct me into your local path for chrome. I have tried to clean this up, but I personally had to reinstall drivers to get it working as well as change the location of the utiles folder. Be prepared for here be dragons!! 😟